+++ src/main/java/step3/GameHistory.java
@@ -6,16 +6,16 @@ import java.util.List;
 import java.util.Map;
 
 public class GameHistory {
-    private final Map<Long, List<Position>> history = new HashMap<>();
+    private final Map<Integer, List<Position>> history = new HashMap<>();
 
-    public void save(Long now, List<Position> tempHistory) {
+    public void save(Integer now, List<Position> tempHistory) {
         List<Position> saveHistory = new ArrayList<>();
         tempHistory.stream()
                 .forEach(saveHistory::add);
         history.put(now, saveHistory);
     }
 
-    public List<Position> getHistory(Long time) {
+    public List<Position> getHistory(Integer time) {
         isValid(time);
         List<Position> saveHistory = history.get(time);
         List<Position> returnHistory = new ArrayList<>();
@@ -24,7 +24,7 @@ public class GameHistory {
         return returnHistory;
     }
 
-    private void isValid(Long time) {
+    private void isValid(Integer time) {
         if (!history.containsKey(time)) {
             throw new IllegalArgumentException("history가 존재하지 않습니다.");
         }
